# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BSYSTEM Operations API
  version: 0.1.0
  description: >
    Fleet-моніторинг серверів BRAVO-Toolkit (backup, LIMS/VetOffice
    maintenance, health/статус служб). Standalone v1 — без інтеграції з
    Integration Core/Module Registry (див. README).
servers:
  - url: /api/v1

components:
  securitySchemes:
    bootstrapSecret:
      type: apiKey
      in: header
      name: X-Bootstrap-Secret
      description: Видається адміністратором Operations разом з увімкненням opt-in на сервері.
    sessionCookie:
      type: apiKey
      in: cookie
      name: ops_session
      description: >
        Etap 3: httpOnly-сесія локального облікового запису, видається
        POST /auth/login. Роль (admin|viewer) визначає доступ —
        approve вимагає admin, решта адмін-маршрутів доступні обом.
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Виданий після approve — reveal-once через GET /enroll/{serverId}.

  schemas:
    ProductType:
      type: string
      enum: [LIMS, VETOFFICE]
    ServerStatus:
      type: string
      enum: [pending, approved, revoked]
    Severity:
      type: string
      enum: [SUCCESS, WARNING, ERROR, CRITICAL]
    EventCategory:
      type: string
      enum: [backup, maintenance, health]
    UserRole:
      type: string
      enum: [admin, viewer]
    ServiceStatusEntry:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
          example: exchangAPI
        status:
          type: string
          enum: [running, stopped, unknown]
    EventPayload:
      type: object
      required: [message]
      properties:
        message:
          type: string
        component:
          type: string
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceStatusEntry'
        details:
          type: object
          additionalProperties: true

paths:
  /auth/login:
    post:
      summary: Вхід локальним обліковим записом (Etap 3). Немає self-registration — акаунти лише через CLI create-admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Вхід успішний, ops_session cookie встановлено
        '400':
          description: Невалідний запит
        '401':
          description: Невірний логін/пароль

  /auth/logout:
    post:
      summary: Вихід — видаляє сесію та cookie
      security:
        - sessionCookie: []
      responses:
        '204':
          description: Вихід виконано
        '401':
          description: Немає активної сесії

  /auth/me:
    get:
      summary: Поточний обліковий запис (username + role)
      security:
        - sessionCookie: []
      responses:
        '200':
          description: username + role
        '401':
          description: Немає активної сесії

  /enroll:
    post:
      summary: Self-enrollment сервера (GUID + bootstrap-секрет)
      security:
        - bootstrapSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serverId, institutionCode, productType, hostname, bootstrapSecret]
              properties:
                serverId:
                  type: string
                  format: uuid
                institutionCode:
                  type: string
                productType:
                  $ref: '#/components/schemas/ProductType'
                hostname:
                  type: string
                bootstrapSecret:
                  type: string
      responses:
        '202':
          description: Прийнято (idempotent — не скидає вже approved/revoked сервер назад у pending)
        '400':
          description: Невалідний запит
        '401':
          description: Невірний bootstrap-секрет
        '503':
          description: Enrollment не сконфігуровано (OPERATIONS_BOOTSTRAP_SECRET відсутній)

  /enroll/{serverId}:
    get:
      summary: Поллінг статусу enrollment; apiKey повертається РІВНО ОДИН РАЗ одразу після approve
      security:
        - bootstrapSecret: []
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Статус (+ apiKey, лише одразу після approve)
        '401':
          description: Невірний bootstrap-секрет
        '404':
          description: Сервер не зареєстрований

  /admin/servers:
    get:
      summary: >
        Список усіх серверів для overview (Etap 3): кожен запис
        збагачено isOnline (з last_heartbeat_at) і latestByCategory
        (останній backup/maintenance/health-запис). Доступно admin і
        viewer.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Масив серверів + heartbeatExpectedIntervalMinutes/heartbeatMissedThreshold
        '401':
          description: Немає активної сесії

  /admin/servers/{serverId}:
    get:
      summary: >
        Деталі сервера для detail-режиму (Etap 3): останні 1000 подій,
        latestByCategory, isOnline. Доступно admin і viewer.
      security:
        - sessionCookie: []
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Сервер, latestByCategory і його події
        '401':
          description: Немає активної сесії
        '404':
          description: Не знайдено

  /admin/servers/{serverId}/approve:
    post:
      summary: Підтвердити pending-сервер, видати API-ключ. Лише роль admin.
      security:
        - sessionCookie: []
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Approved
        '401':
          description: Немає активної сесії
        '403':
          description: Сесія є, але роль не admin
        '404':
          description: Не знайдено
        '409':
          description: Сервер не у статусі pending

  /events:
    post:
      summary: >
        Прийом подій від агента (Archive/BazaSync/Maintenance/Health) —
        ЗАВЖДИ, незалежно від NotificationMode Discord-провайдера
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, severity, payload]
              properties:
                category:
                  $ref: '#/components/schemas/EventCategory'
                severity:
                  $ref: '#/components/schemas/Severity'
                payload:
                  $ref: '#/components/schemas/EventPayload'
      responses:
        '202':
          description: Прийнято
        '400':
          description: Невалідний payload
        '401':
          description: Невірний/відкликаний API-ключ

  /heartbeat:
    post:
      summary: Періодичний heartbeat (незалежно від подій) — джерело для offline-детекції (Etap 4)
      security:
        - apiKey: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                bravoVersion:
                  type: string
      responses:
        '202':
          description: Прийнято
        '401':
          description: Невірний/відкликаний API-ключ